Agriculture Plough
Agriculture Plough is a primary tillage implement which is used to open the new fields and to process the stony areas. Domestic Disc Plough can be used easily at rocky and rooted areas. The field after plowing with disk plow result in large soil clods and uneven topography. The disc plough is designed to work in all types of soil for functions such as soil mixing, soil raising, soil turning and soil breaking.